Responsive Congress Website using Bootstrap
This is a web page that provides an interface to perform congress search
using Sunlight Congress APIs.Ajax and JSON is used to communicate
between the server and the client. Bootstrap for responsiveness.
AngularJS, CSS and HTML for the overall structure and design of the web
page. JavaScript is used for client-side scripting. PHP on the
server-side.